+++ test/com/arpnetworking/metrics/portal/scheduling/JobSchedulerTest.java
@@ -21,9 +21,10 @@ import akka.testkit.javadsl.TestKit;
 import com.arpnetworking.commons.java.time.ManualClock;
 import com.arpnetworking.metrics.portal.AkkaClusteringConfigFactory;
 import com.arpnetworking.metrics.portal.scheduling.impl.MapJobRepository;
-import com.arpnetworking.metrics.portal.scheduling.impl.OneOffSchedule;
 import com.arpnetworking.metrics.portal.scheduling.impl.PeriodicSchedule;
 import com.typesafe.config.ConfigFactory;
+import models.internal.scheduling.Job;
+import models.internal.scheduling.Schedule;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
